Implement Gamedistribution SDK.
Game ID is required in the index.html file, so we use Stencyl 4.1.0 beta's per-extension game data features to pass that in.
1. gdGameId in game-data.xml and stencyl-addons.xml
2. Set to GD_GAME_ID in include.xml
3. Value is accessed as ::GD_GAME_ID:: in index.html

The sdk is added as a js dependency in include.xml. This causes the SDK to be loaded before the game starts, so we receive a few SDK events before Stencyl's ready for it. So we use an event queue in index.html to hold events until we're ready for them.